requirements.txt
torch
torchvision
gradio
gdown
Pillow
opencv-python
numpy